dcos/nodeutil: add headers to mesos-id request
this is non-breaking change that adds ability for user to pass custom request
headers to `MesosID` function.

This will be used for authorization. Instead of system account, we will make
a request on user's behalf. In order to do so, we need to be able to pass
Authorization header to http request.