dcos/nodeutil: remove .Role() method
detecting a node's role in runtime is not very trivial task and leads to
errors and bugs in software, since we cannot predict users environment.

A source of truth for node's role is gen library. Every service which is using
this package must be aware of the node's role from it's config file, which
in turn, is generated by dcos-config.yaml. This approach will eliminate the
need of hardcoding files which describe the roles and will make the library
robust.

related issues: https://mesosphere.atlassian.net/browse/DCOS-11680