Add elector type.
The elector, once started, will emit events on the Events()
channel to signify when leadership changes, or when errors occur.

Additionally, the current leader of the group (defined as
having the same basePath) may be queried at any time using the
LeaderIdent() function.

If any error is received no the Events() channel, the elector
should no longer be considered valid and the client should
likely restart.